+++ magentoerpconnect_pricing/product.py
@@ -21,7 +21,7 @@
 
 
 from openerp.tools.translate import _
-from openerp.addons.connector.queue.job import job
+from openerp.addons.connector.queue.job import job, related_action
 from openerp.addons.connector.exception import FailedJobError
 from openerp.addons.connector.unit.mapper import (mapping,
                                                   only_create
@@ -32,8 +32,12 @@ from openerp.addons.magentoerpconnect.unit.export_synchronizer import (
 from openerp.addons.magentoerpconnect.backend import magento
 from openerp.addons.magentoerpconnect import product
 from openerp.addons.magentoerpconnect.connector import get_environment
+from openerp.addons.magentoerpconnect.related_action import (
+    unwrap_binding,
+)
 
 
+# TODO: replace a price mapper only, not the full mapper
 @magento(replacing=product.ProductImportMapper)
 class ProductImportMapper(product.ProductImportMapper):
     _model_name = 'magento.product.product'
@@ -142,6 +146,7 @@ def product_price_changed(session, model_name, record_id, fields=None):
 
 
 @job
+@related_action(action=unwrap_binding)
 def export_product_price(session, model_name, record_id, website_id=None):
     """ Export the price of a product. """
     product_bind = session.browse(model_name, record_id)